added test for validate_coordinates.py
